refactor(hooks): split hooks.rs into hooks/ submodule (#103)
Split src/hooks.rs (1240 lines) into 6 files under src/hooks/ to
clarify responsibility boundaries that had drifted from the design
intent in CLAUDE.md.

- mod.rs        — public API re-export (3 functions)
- executor.rs   — execute_hooks_* family + execute_hooks_impl
- runner.rs     — execute_command + tail buffer + spinner UI
- files.rs      — pattern expansion + file copy
- symlink.rs    — symlink idempotency
- output.rs     — emit_line + format_duration

Production public API surface narrowed from 5 to 3 (emit_line,
execute_hooks_with_mp, execute_hooks_lenient_with_mp). Deleted
HookExecutor trait + RealHookExecutor + MockHookExecutor (production
wiring removed in #102; reattachment was infeasible). Demoted dead
pub fn execute_hooks (line 156) to #[cfg(test)] pub(super), as both
its callers were tests.

Tests: redistributed 27 tests by responsibility area into
per-submodule mod tests blocks (runner 3 / files 12 / symlink 7 /
executor 5). Mock tests (2) were deleted alongside the trait.

Pure refactor — zero behavior change. Callers (commands/{add,create,
rm,sync}.rs) untouched.
